Rename the ca module to automata
