+++ docs/source/automata.rst
@@ -1,0 +1,31 @@
+Cellular Automata
+=================
+
+API Documentation
+-----------------
+
+.. automodule:: neet.automata
+
+Elementary Cellular Automata
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+    
+    .. autoclass:: neet.automata.ECA       
+    
+Attributes
+""""""""""
+        
+        .. autoattribute:: neet.automata.ECA.code
+             
+        .. autoattribute:: neet.automata.ECA.boundary
+    
+Initialization
+""""""""""""""
+        
+        .. automethod:: neet.automata.ECA.__init__
+
+Methods
+"""""""
+        
+        .. automethod:: neet.automata.ECA.update
+        
+        .. automethod:: neet.automata.ECA.step
