--- docs/source/cas.rst
@@ -1,31 +1,0 @@
-Cellular Automata
-=================
-
-API Documentation
------------------
-
-.. automodule:: neet.ca
-
-Elementary Cellular Automata
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-    
-    .. autoclass:: neet.ca.ECA       
-    
-Attributes
-""""""""""
-        
-        .. autoattribute:: neet.ca.ECA.code
-             
-        .. autoattribute:: neet.ca.ECA.boundary
-    
-Initialization
-""""""""""""""
-        
-        .. automethod:: neet.ca.ECA.__init__
-
-Methods
-"""""""
-        
-        .. automethod:: neet.ca.ECA.update
-        
-        .. automethod:: neet.ca.ECA.step
