+++ neet/automata.py
